requirements.txt
mysql-connector-python
cryptography
requests
pyjwt
flask